Move ISV Debugger activation steps to salesforcedx-vscode-apex-debugger (#889)
@W-5702036@
