+++ packages/salesforcedx-vscode-apex/package.json
@@ -21,7 +21,17 @@
   "engines": {
     "vscode": "^1.26.0"
   },
-  "categories": ["Languages"],
+  "categories": ["Programming Languages"],
+  "dependencies": {
+    "@salesforce/apex-tmlanguage": "1.1.0",
+    "@salesforce/salesforcedx-utils-vscode": "44.18.0",
+    "expand-home-dir": "0.0.3",
+    "find-java-home": "0.2.0",
+    "path-exists": "3.0.0",
+    "shelljs": "^0.8.1",
+    "vscode-extension-telemetry": "0.0.17",
+    "vscode-languageclient": "5.1.1"
+  },
   "devDependencies": {
     "@types/chai": "^4.0.0",
     "@types/mocha": "2.2.38",
@@ -214,15 +224,5 @@
         "path": "./snippets/apex.json"
       }
     ]
-  },
-  "dependencies": {
-    "@salesforce/apex-tmlanguage": "1.1.0",
-    "@salesforce/salesforcedx-utils-vscode": "44.18.0",
-    "expand-home-dir": "0.0.3",
-    "find-java-home": "0.2.0",
-    "path-exists": "3.0.0",
-    "shelljs": "^0.8.1",
-    "vscode-extension-telemetry": "0.0.17",
-    "vscode-languageclient": "5.1.1"
   }
 }
